% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpscarf_add_height.R
\name{alpscarf_add_height}
\alias{alpscarf_add_height}
\title{calculate conformity/revisiting scores and translate them into bar height}
\usage{
alpscarf_add_height(dwell_df = NULL, expected_order = NULL,
  height_mode = c("linear", "exponential"), scale_factor = 0.1,
  base_factor = 2)
}
\arguments{
\item{dwell_df}{a data frame of AOI visits, contains at least 2 columns, "p_name" "AOI"}

\item{expected_order}{expected visit order, two columns "AOI" and "AOI_order"}

\item{height_mode}{to select between linear mode and exponential mode of mountain height}

\item{scale_factor}{to specify scale of mountain height}

\item{base_factor}{to specify the base of exponent which changes mountain shape, only valid in exponential mode}
}
\value{
a tibble with conformity/revisiting scores and bar height information
}
\description{
calculate conformity/revisiting scores and translate them into bar height
}
