% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpscarf_add_width.R
\name{alpscarf_add_width}
\alias{alpscarf_add_width}
\title{translate durations into bar width}
\usage{
alpscarf_add_width(dwell_df = NULL)
}
\arguments{
\item{dwell_df}{a data frame contains at least 3 columns, "p_name" "AOI" "dwell_duration"}
}
\value{
a tibble with bar width information
}
\description{
translate durations into bar width
}
