% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpscarf_statistics.R
\name{alpscarf_calculate_statistics}
\alias{alpscarf_calculate_statistics}
\title{Calculate statistics (order_following/rereading ratio, Levenshtein/Jaro distance) assist reading Alpscarf}
\usage{
alpscarf_calculate_statistics(df_alp_p, aoi_names_pages_seq)
}
\arguments{
\item{df_alp_p}{dataset with Alpscarf info, at least four columns "p_name" "AOI" "conformity_score" "revisiting_score"}

\item{aoi_names_pages_seq}{expected visit order, two columns "AOI" and "AOI_order"}
}
\value{
states (order_following/rereading ratio, Levenshtein/Jaro distance)
}
\description{
Calculate statistics (order_following/rereading ratio, Levenshtein/Jaro distance) assist reading Alpscarf
}
