% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpscarf_height_trans.R
\name{alpscarf_conform}
\alias{alpscarf_conform}
\title{Calculation of conformity score}
\usage{
alpscarf_conform(df_p, aoi_names_pages_seq, s_min = 2)
}
\arguments{
\item{df_p}{dataset contains AOI visits}

\item{aoi_names_pages_seq}{expected visit order, two columns "AOI" and "AOI_order"}

\item{s_min}{shortest length of sequence to consider}
}
\value{
conformity scores
}
\description{
Calculation of conformity score
}
