% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpscarf_plot.R
\name{alpscarf_plot}
\alias{alpscarf_plot}
\title{Alpscarf plot generation}
\usage{
alpscarf_plot(alp_df = NULL, palette = NULL, focus_mode = c("transition",
  "duration"), ymax = 4, plot_type = c("alpscarf", "scarf"),
  creek_offset = 0.04, creek_size = 2, title = NULL,
  NORMALIZED_VIEW = TRUE, max_nr_transitions = 0,
  max_sum_dwell_duration_log = 0)
}
\arguments{
\item{alp_df}{a data frame containing the bar width/height info of Alpscarf, contains at least 6 columns "AOI" "trial" "bar_position" "dwell_duration_log" "seq_bar_length" "re_reading_bar_length"}

\item{palette}{the color definition}

\item{focus_mode}{to select between transition-focus mode and duration-focus mode}

\item{ymax}{the range (in y-axis) of plot}

\item{plot_type}{to select between Alpscarf plot and traditional scarf plot}

\item{creek_offset}{to adjust the position of creeks, default = 0.04}

\item{creek_size}{to adjust the size of creeks, default = 2}

\item{title}{print title; e.g., participant name}

\item{NORMALIZED_VIEW}{to select between normalized view (TRUE) and unnormalized view (FALSE), default = TRUE. In normalized view, all generated Alpscarf are in the same length. In unnormalized view, users need to specify the maximum length of generated Alpscarf.}

\item{max_nr_transitions}{specify the maximum number of transitions in given dataset. Only used in unnormalized view (NORMALIZED_VIEW= FALSE).}

\item{max_sum_dwell_duration_log}{specify the maximum number of the sum of total dwell durations (in log scale) in given dataset. Only used in unnormalized view. (NORMALIZED_VIEW= FALSE)}
}
\value{
Alpscarf plot as a ggplot2 object
}
\description{
Alpscarf plot generation
}
