% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpscarf_plot_gen.R
\name{alpscarf_plot_gen}
\alias{alpscarf_plot_gen}
\title{Alpscarf plot generation}
\usage{
alpscarf_plot_gen(df_p = NULL, my_palette = NULL, TRANSITION_FOCUS = TRUE,
  max_y_value = 4, ALPSCARF_EN = TRUE, creek_offset = 0.04,
  creek_size = 2, TITLE_EN = TRUE, NORMALIZED_VIEW = TRUE,
  max_nr_transitions = 0, max_sum_dwell_duration_category = 0)
}
\arguments{
\item{df_p}{dataset contains the bar width/height info of Alpscarf, contains at least 6 columns "AOI" "trial" "bar_position" "dwell_duration_category" "seq_bar_length" "re_reading_bar_length"}

\item{my_palette}{the color definition}

\item{TRANSITION_FOCUS}{to select between transition-focus (TRUE) and duration-focus (FALSE) mode, default = TRUE}

\item{max_y_value}{the range (in y-axis) to plot}

\item{ALPSCARF_EN}{to select between Alpscarf plot (TRUE) or traditional scarf plot (FALSE), default = TRUE}

\item{creek_offset}{to adjust the position of creeks, default = 0.04}

\item{creek_size}{to adjust the size of creeks, default = 2}

\item{TITLE_EN}{print participant name, default = TRUE}

\item{NORMALIZED_VIEW}{to select between normalized view (TRUE) and unnormalized view (FALSE), default = TRUE. In normalized view, all generated Alpscarf are in the same length. In unnormalized view, users need to specify the maximum length of generated Alpscarf.}

\item{max_nr_transitions}{specify the maximum number of transitions in given dataset. Only used in unnormalized view (NORMALIZED_VIEW= FALSE).}

\item{max_sum_dwell_duration_category}{specify the maximum number of the sum of total dwell durations (in log scale) in given dataset. Only used in unnormalized view. (NORMALIZED_VIEW= FALSE)}
}
\value{
Alpscarf plot
}
\description{
Alpscarf plot generation
}
