% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpscarf_height_trans.R
\name{alpscarf_revisit}
\alias{alpscarf_revisit}
\title{Calculation of revisiting score}
\usage{
alpscarf_revisit(df_p, w = 3)
}
\arguments{
\item{df_p}{dataset contains AOI visits}

\item{w}{sequence length of revisir type to consider}
}
\value{
revisiting scores
}
\description{
Calculation of revisiting score
}
