% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpscarf_statistics.R
\name{alpscarf_stats}
\alias{alpscarf_stats}
\title{Calculate statistics (order_following/rereading ratio, Levenshtein/Jaro distance) to assist in reading Alpscarf}
\usage{
alpscarf_stats(alp_df = NULL, expected_order = NULL)
}
\arguments{
\item{alp_df}{a data frame with Alpscarf info, at least four columns "p_name" "AOI" "conformity_score" "revisiting_score"}

\item{expected_order}{expected visit order, two columns "AOI" and "AOI_order"}
}
\value{
stats (order_following/rereading ratio, Levenshtein/Jaro distance)
}
\description{
Calculate statistics (order_following/rereading ratio, Levenshtein/Jaro distance) to assist in reading Alpscarf
}
