% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpscarf_width_trans.R
\name{alpscarf_width_trans}
\alias{alpscarf_width_trans}
\title{translate durations into bar width}
\usage{
alpscarf_width_trans(dwell_df = NULL)
}
\arguments{
\item{dwell_df}{dataset contains at least 3 columns, "p_name" "AOI" "dwell_duration"}
}
\value{
dataset with bar width information
}
\description{
translate durations into bar width
}
