% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aoi_names_pages_seq}
\alias{aoi_names_pages_seq}
\title{Expected AOI visit order}
\format{A data frame with 37 rows and 3 variables:
\describe{
  \item{AOI}{the name of AOI}
  \item{AOI_order}{the expected visit order of AOIs; must be continuous integers}
  \item{color}{a set of color definitions in HEX code; a 1-to-1 mapping to AOI_order}
}}
\usage{
aoi_names_pages_seq
}
\description{
A dataset containing the AOI names, their expected visit order, and color association. The variables are as follows:
}
\keyword{datasets}
