% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eye_movement_data_systhetic}
\alias{eye_movement_data_systhetic}
\title{AOI visits (dwells)}
\format{A data frame with 568 rows and 3 variables:
\describe{
  \item{p_name}{participant name (P1--P3)}
  \item{AOI}{the visited AOI of one dwell}
  \item{dwell_duration}{the total dwell time of one dwell}
}}
\usage{
eye_movement_data_systhetic
}
\description{
A synthetic eye-movement dataset containing a sequence of AOI visits for multiple participants. The variables are as follows:
}
\keyword{datasets}
