% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsa_reduced_dwell_df}
\alias{lsa_reduced_dwell_df}
\title{AOI visits (dwells)}
\format{A data frame with 3048 rows and 3 variables:
\describe{
  \item{p_name}{participant name (P1--P18)}
  \item{AOI}{the visited AOI of one dwell}
  \item{dwell_duration}{the total dwell time of one dwell}
}}
\usage{
lsa_reduced_dwell_df
}
\description{
A dataset containing a sequence of AOI visits for multiple participants. The variables are as follows:
}
\keyword{datasets}
