% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpscarf_plot_gen.R
\name{alpscarf_plot_gen}
\alias{alpscarf_plot_gen}
\title{Alpscarf plot generation}
\usage{
alpscarf_plot_gen(df_p = NULL, my_palette = NULL, TRANSITION_FOCUS = TRUE,
  max_y_value = 4, ALPSCARF_EN = TRUE, creek_offset = 0.04,
  creek_size = 2, TITLE_EN = TRUE)
}
\arguments{
\item{df_p}{a data frame containing the bar width/height info of Alpscarf, contains at least 6 columns "AOI" "trial" "bar_position" "dwell_duration_category" "seq_bar_length" "re_reading_bar_length"}

\item{my_palette}{the color definition}

\item{TRANSITION_FOCUS}{to select between transition-focus (TRUE) and duration-focus (FALSE) mode, default = TRUE}

\item{max_y_value}{the range (in y-axis) to plot}

\item{ALPSCARF_EN}{to select between Alpscarf plot (TRUE) or traditional scarf plot (FALSE), default = TRUE}

\item{creek_offset}{to adjust the position of creeks, default = 0.04}

\item{creek_size}{to adjust the size of creeks, default = 2}

\item{TITLE_EN}{print participant name, default = TRUE}
}
\value{
Alpscarf plot as a ggplot2 object
}
\description{
Alpscarf plot generation
}
