% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpscarf_width_trans.R
\name{alpscarf_width_trans}
\alias{alpscarf_width_trans}
\title{translate durations into bar width info of Alpscarf}
\usage{
alpscarf_width_trans(df_p)
}
\arguments{
\item{df_p}{dataset contains 4 columns "AOI" "dwell_duration" "re_reading_bar_length" "seq_bar_length"}
}
\value{
translated bar width/height info
}
\description{
translate durations into bar width info of Alpscarf
}
